Rearrange embargo exceptions
