+++ dandiapi/api/services/dandiset/__init__.py
@@ -4,10 +4,8 @@ from django.db import transaction
 
 from dandiapi.api.models.dandiset import Dandiset
 from dandiapi.api.models.version import Version
-from dandiapi.api.services.dandiset.exceptions import (
-    DandisetAlreadyExistsError,
-    DandisetUnEmbargoInProgressError,
-)
+from dandiapi.api.services.dandiset.exceptions import DandisetAlreadyExistsError
+from dandiapi.api.services.embargo.exceptions import DandisetUnEmbargoInProgressError
 from dandiapi.api.services.exceptions import AdminOnlyOperationError, NotAllowedError
 from dandiapi.api.services.version.metadata import _normalize_version_metadata
