+++ dandiapi/api/services/dandiset/exceptions.py
@@ -7,15 +7,3 @@ from dandiapi.api.services.exceptions import DandiError
 
 class DandisetAlreadyExistsError(DandiError):
     http_status_code = status.HTTP_400_BAD_REQUEST
-
-
-class UnauthorizedEmbargoAccessError(DandiError):
-    http_status_code = status.HTTP_401_UNAUTHORIZED
-    message = (
-        'Authentication credentials must be provided when attempting to access embargoed dandisets'
-    )
-
-
-class DandisetUnEmbargoInProgressError(DandiError):
-    http_status_code = status.HTTP_400_BAD_REQUEST
-    message = 'Dandiset modification not allowed during un-embargo'
