+++ dandiapi/api/services/embargo/exceptions.py
@@ -18,3 +18,15 @@ class DandisetNotEmbargoedError(DandiError):
 class DandisetActiveUploadsError(DandiError):
     http_status_code = status.HTTP_400_BAD_REQUEST
     message = 'Dandiset un-embargo not allowed with active uploads'
+
+
+class DandisetUnEmbargoInProgressError(DandiError):
+    http_status_code = status.HTTP_400_BAD_REQUEST
+    message = 'Dandiset modification not allowed during un-embargo'
+
+
+class UnauthorizedEmbargoAccessError(DandiError):
+    http_status_code = status.HTTP_401_UNAUTHORIZED
+    message = (
+        'Authentication credentials must be provided when attempting to access embargoed dandisets'
+    )
